[ALPS02188337] MMProfile: shrink kernel log in mmprofile
[Detail] need to shrink kernel log for mmprofile driver.
Depending on specific build, it may cause "Kernel API dump"
exception for mmprofile driver had print too much log. The log was
printed by calling the trace functions in MMProfileLog_Int().
    fix build error -
     /worktmp/src_alps-trunk-m0.basic_kernel/kernel-3.18/drivers/misc/
     mediatek/mmp/src/mmprofile.c:1293:3: error: passing argument 3 of
     kstrtoul from incompatible pointer type [-Werror]
         if (0 == kstrtoul(p, 10, &value) && 0 != value)

[Solution] The following changes are made to shrink kernel log:
1. adjust the definition and pre-compile marco, which can disables
the trace functions for mmprofile driver.
2. add mmprofile_write implementation, which enables user to
dynamically set "mmp_log_on" and "mmp_trace_log_on" which controls
mmprofile driver log output.

[Feature] Others

MTK-Commit-Id: e95212891c4db94e4408773ba8ab0274c5b73ac1

CR-Id: ALPS02188337

+++ drivers/misc/mediatek/mmp/src/mmprofile.c
@@ -1279,7 +1279,7 @@ static void process_dbg_cmd(char *cmd)
 {
 	if (0 == strncmp(cmd, "mmp_log_on:", 11)) {
 		char *p = (char *)cmd + 11;
-		int value;
+		unsigned long value;
 
 		if (0 == kstrtoul(p, 10, &value) && 0 != value)
 			mmp_log_on = 1;
@@ -1288,7 +1288,7 @@ static void process_dbg_cmd(char *cmd)
 		MMP_MSG("mmp_log_on=%d\n", mmp_log_on);
 	} else if (0 == strncmp(cmd, "mmp_trace_log_on:", 17)) {
 		char *p = (char *)cmd + 17;
-		int value;
+		unsigned long value;
 
 		if (0 == kstrtoul(p, 10, &value) && 0 != value)
 			mmp_trace_log_on = 1;
